This CL merges the following CLs from aosp/idea133 into aosp/master.

dabd634b Snapshot e2750ea61032f2a041cb012bb7b90cffa0deba73 from idea/133.124 of git://git.jetbrains.org/idea/community.git
29ab773f Make default test target run Android specific tests
967117a7 Merge "Make default test target run Android specific tests" into idea133
1eb71f20 Bump memory settings for unit tests
b13ea0d4 Don't bind the delete key to restoring property in layout editor
a20ccfa9 Add Gradle import module
0980f4a2 Add gradle-import to build script
cee6f8ca Fix Gradle notification lookup on Windows.
8668e1be Snapshot 020d29497847701e84e383d965abf543b80758e2 from idea/133.370 of git://git.jetbrains.org/idea/community.git
18f77669 Merge remote-tracking branch 'aosp/snapshot-master' into merge
36ac8cba Update from Guava 13.0.1 to Guava 15
4d451f93 Update libraries to lombok 0.2.2
ee06b1d0 Remove lint-cli dependency from the idea project
ab73dade Updater: add unit tests.
dd558b6e Updater: on Windows, add "Retry" on file op failures.
f2f7178a Snapshot c11f3ac9bbde3f85d1f837ec3eb48a395ed7dd10 from idea/133.471 of git://git.jetbrains.org/idea/community.git
5e4c77db Merge remote-tracking branch 'aosp/snapshot-master' into merge
8d957349 Fix junit.jar path in updater project.
58c3e0ae Include cloud tools tests in default test group
63cd9779 Temporarily disable errors in project structure dialog
e2d6089d Snapshot b9931c55d2175b6552f90b2225eb09c13bd6dfed from idea/133.609 of git://git.jetbrains.org/idea/community.git
031a291e Merge remote-tracking branch 'aosp/snapshot-master' into merge
ea628d6e Remove versions from Info.plist
809cb3e7 Snapshot 9e6329d622cc9649c9c035f28faddc29564a5b7a from idea/133.696 of git://git.jetbrains.org/idea/community.git
d6cfe6ec Merge remote-tracking branch 'aosp/snapshot-master' into merge
38f8c6f0 Gracefully handle build.gradle files in modules without a configured JDK
70ae6f2a Snapshot dc1944e804515a346297e368c3b9c35a203c9912 from idea/133.818 of git://git.jetbrains.org/idea/community.git
ac91a6de Merge remote-tracking branch 'aosp/snapshot-master' into merge
2f51d957 Gradle: respect build classpath order and use both classes and sources jars if available
0ecdb509 Snapshot c50a8ad26a72432f26e39046d6a6f21fd7a190ee from idea/134.1160 of git://git.jetbrains.org/idea/community.git
e8c22ad7 Merge remote-tracking branch 'aosp/snapshot-master' into merge
72253f7d Turn off android framework detection
93f77ee6 Temporarily remove GCT tests
88f318c9 Snapshot 34f078c3452e79ba209d28a551962857e0970e5d from idea/134.1342 of git://git.jetbrains.org/idea/community.git
afb54e4b Merge remote-tracking branch 'aosp/snapshot-master' into merge
4dc795dc Fix updater UI tests.
57a49ed1 Studio patch: more logging.
aa614ee0 table greyer (ability to disable a table)
5c571417 Use Gradle model prebuilts v0.9.0 in Studio.
cb38c25d build script: Run jarjar on the updater
f273ca07 Add App Engine templates dir to build
7607404f Removed android-builder library from Studio (not needed.)
8f29b4eb Merge idea133 changes into master

+++ platform/platform-api/src/com/intellij/execution/process/CapturingProcessHandler.java
@@ -18,6 +18,7 @@ package com.intellij.execution.process;
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.configurations.GeneralCommandLine;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.progress.ProgressIndicator;
 import org.jetbrains.annotations.NotNull;
 
 import java.nio.charset.Charset;
@@ -30,10 +31,10 @@ import java.nio.charset.Charset;
 public class CapturingProcessHandler extends OSProcessHandler {
   private static final Logger LOG = Logger.getInstance(CapturingProcessHandler.class);
   private final ProcessOutput myOutput = new ProcessOutput();
-
+  
   public CapturingProcessHandler(@NotNull GeneralCommandLine commandLine) throws ExecutionException {
     super(commandLine);
-    addProcessListener(new CapturingProcessAdapter(myOutput));
+    addProcessListener(createProcessAdapter(myOutput));
   }
 
   public CapturingProcessHandler(final Process process) {
@@ -46,9 +47,13 @@ public class CapturingProcessHandler extends OSProcessHandler {
 
   public CapturingProcessHandler(final Process process, final Charset charset, final String commandLine) {
     super(process, commandLine, charset);
-    addProcessListener(new CapturingProcessAdapter(myOutput));
+    addProcessListener(createProcessAdapter(myOutput));
   }
 
+  protected CapturingProcessAdapter createProcessAdapter(ProcessOutput processOutput) {
+    return new CapturingProcessAdapter(processOutput);
+  }
+  
   public ProcessOutput runProcess() {
     startNotify();
     if (waitFor()) {
@@ -66,6 +71,16 @@ public class CapturingProcessHandler extends OSProcessHandler {
    * @param timeoutInMilliseconds non-positive means infinity
    */
   public ProcessOutput runProcess(int timeoutInMilliseconds) {
+    return runProcess(timeoutInMilliseconds, true);
+  }
+
+  /**
+   * Starts process with specified timeout
+   *
+   * @param timeoutInMilliseconds non-positive means infinity
+   * @param destroyOnTimeout whether to kill the process after timeout passes
+   */
+  public ProcessOutput runProcess(int timeoutInMilliseconds, boolean destroyOnTimeout) {
     if (timeoutInMilliseconds <= 0) {
       return runProcess();
     }
@@ -75,7 +90,9 @@ public class CapturingProcessHandler extends OSProcessHandler {
         myOutput.setExitCode(getProcess().exitValue());
       }
       else {
-        destroyProcess();
+        if (destroyOnTimeout) {
+          destroyProcess();
+        }
         myOutput.setTimeout();
       }
       return myOutput;
@@ -88,5 +105,25 @@ public class CapturingProcessHandler extends OSProcessHandler {
       return myCharset;
     }
     return super.getCharset();
+  }
+
+  @NotNull
+  public ProcessOutput runProcessWithProgressIndicator(@NotNull ProgressIndicator indicator) {
+    startNotify();
+    while (!waitFor(100)) {
+      if (indicator.isCanceled()) {
+        if (!isProcessTerminating() && !isProcessTerminated()) {
+          destroyProcess();
+        }
+        break;
+      }
+    }
+    if (waitFor()) {
+      myOutput.setExitCode(getProcess().exitValue());
+    }
+    else {
+      LOG.info("runProcess: exit value unavailable");
+    }
+    return myOutput;
   }
 }
